images/requirements.txt
ultralytics
opencv-python
pytesseract
Pillow
pandas
